+++ services/iaas/CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.24.0 (2025-06-05)
+- **Feature:** Add waiters for async operations: `CreateBackupWaitHandler`, `DeleteBackupWaitHandler`, `RestoreBackupWaitHandler`
+- **Feature:** Add Waiters for async operations: `CreateSnapshotWaitHandler`, `DeleteSnapshotWaitHandler`
+
 ## v0.23.0 (2025-05-15)
 - **Breaking change:** Introduce interfaces for `APIClient` and the request structs